Update codequality job template to use locked image
closes #45

Like we did for gitlab-org/gitlab-ci-yml!138, we are now using a local
image on our registries, and the version is mapping to the current
GitLab release (minor).

I renamed SAST_VERSION to SCA_VERSION because both jobs are probably
going to be merged in the future (they do almost the same things).