Maven.gitlab-ci: specify absolute path to .m2/
The existing code is working, but breaks as soon as mvn is run from a
directory other than the root, e.g.

  script:
    - cd subdir && mvn test

To be candidate for caching, .m2 must be within the project directory,
but we can specify its absolute location using $CI_PROJECT_DIR/.m2/.
It still works in the base case, and also works in the "cd subdir"
case above hence won't create surprise when users try to adapt the
template to their needs.